Moved html helper extensions to the same file
